scheduler: `RescheduleTracker` dropped if follow-up fail placements
When an allocation fails it triggers an evaluation. The evaluation is
processed and the scheduler sees it needs to reschedule, which
triggers a follow-up eval. The follow-up eval creates a plan to
`(stop 1) (place 1)`. The replacement alloc has a `RescheduleTracker`
(or gets its `RescheduleTracker` updated).

But in the case where the follow-up eval can't place all allocs (there
aren't enough resources), it can create a partial plan to
`(stop 1) (place 0)`. It then creates a blocked eval. The plan applier
stops the failed alloc. Then when the blocked eval is processed, the
job is missing an allocation, so the scheduler creates a new
allocation. This allocation is _not_ a replacement from the
perspective of the scheduler, so it's not handed off a
`RescheduleTracker`.